+++ packages/vscode-extension/src/commands/index.ts
@@ -8,3 +8,4 @@ export const commandContributions: CommandContribution[] = [];
 export const RUN_PIPELINE_COMMAND = 'nextflowIde.runPipeline';
 export const RESUME_RUN_COMMAND = 'nextflowIde.resumeRun';
 export const STOP_RUN_COMMAND = 'nextflowIde.stopRun';
+export const SHOW_RUN_DETAILS_COMMAND = 'nextflowIde.showRunDetails';